refactor: WinningNumberFactory 추가 및 검증위치 변경
내용:
- Factory 를 통해서 WinningNumber 를 생성한다
- 기존에 WinningNumber 에 수행하던 검증기능을 Factory
에서 수행한다
