+++ src/main/java/lotto/Application.java
@@ -8,7 +8,7 @@ public class Application {
     public static void main(String[] args) {
         LottoController lottoController = AppConfig.lottoController();
         lottoController.createLottoNumbers();
-        lottoController.createLotto();
-        lottoController.startLotto();
+        lottoController.createLottoNumberMatcher();
+        lottoController.startLottoGame();
     }
 }
