+++ src/main/java/lotto/domain/winningnumber/factory/WinningNumberFactory.java
@@ -1,0 +1,8 @@
+package lotto.domain.winningnumber.factory;
+
+import lotto.domain.winningnumber.WinningNumber;
+
+public interface WinningNumberFactory {
+
+    WinningNumber createWinningNumber();
+}
