+++ src/main/java/lotto/domain/winningnumber/factory/validation/CountValidator.java
@@ -1,4 +1,4 @@
-package lotto.domain.winningnumber.validation;
+package lotto.domain.winningnumber.factory.validation;
 
 import java.util.Set;
