+++ src/test/java/lotto/domain/winningnumber/factory/validation/CountValidatorTest.java
@@ -1,6 +1,6 @@
-package lotto.domain.winningnumber.validation;
+package lotto.domain.winningnumber.factory.validation;
 
-import static lotto.domain.winningnumber.validation.WinningNumberValidator.ERROR_RANGE_MESSAGE;
+import static lotto.domain.winningnumber.factory.validation.WinningNumberValidator.ERROR_COUNT_MESSAGE;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
@@ -9,12 +9,10 @@ import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
-class RangeValidatorTest {
+class CountValidatorTest {
 
-    WinningNumberValidator validator = new RangeValidator();
+    WinningNumberValidator validator = new CountValidator();
     Set<String> winningNumbers = new HashSet<>();
 
     @BeforeEach
@@ -26,19 +24,17 @@ class RangeValidatorTest {
         winningNumbers.add("5");
     }
 
-    @ParameterizedTest
-    @DisplayName("1~45 사이 값이 아니면 EX 발생")
-    @ValueSource(strings = {"46", "0", "-1"})
-    void range_ex(String ex) {
-        winningNumbers.add(ex);
+    @Test
+    @DisplayName("당첨번호의 갯수가 6개가 아니면 EX 발생")
+    void count_ex() {
         assertThatIllegalArgumentException().isThrownBy(() -> validator.validate(winningNumbers))
-                .withMessageContaining(ERROR_RANGE_MESSAGE);
+                .withMessageContaining(ERROR_COUNT_MESSAGE);
     }
 
     @Test
-    @DisplayName("1~45 사이 값이면 통과")
-    void range_success() {
+    @DisplayName("당첨번호의 갯수가 6개면 통과")
+    void count_success() {
+        winningNumbers.add("6");
         assertThatNoException().isThrownBy(() -> validator.validate(winningNumbers));
-
     }
 }
